+++ src/node.ts
@@ -2,24 +2,6 @@
  * ASTノード
 */
 
-export function NODE(type: string, params: Record<string, any>, loc: { column: number, line: number }): Node {
-	const node: Record<string, any> = { type };
-	for (const key of Object.keys(params)) {
-		if (params[key] !== undefined) {
-			node[key] = params[key];
-		}
-	}
-	node.loc = loc;
-	return node as Node;
-}
-
-export function CALL_NODE(name: string, args: Node[], loc: { column: number, line: number }): Node {
-	return NODE('call', {
-		target: NODE('identifier', { name }, loc),
-		args,
-	}, loc);
-}
-
 export type Loc = {
 	line: number;
 	column: number;
