+++ src/parser/syntaxes/toplevel.ts
@@ -1,4 +1,4 @@
-import { NODE } from '../../node.js';
+import { NODE } from '../utils.js';
 import { TokenKind } from '../token.js';
 import { AiScriptSyntaxError } from '../../error.js';
 import { parseDefStatement, parseStatement } from './statements.js';
